Remove the grid layout
Arguably, the grid is not a good fit. We're working with a single stream UX as opposed to several sessions happening at once, and it just makes way more sense that all the data is presented in a single column.

It makes the presentation slightly worse on wide screens, but also makes it significantly less confusing in re. overlapping times.

Signed-off-by: Vladimir Pouzanov <farcaller@gmail.com>